split stub and arginfo
