+++ src/Mix_Chunk.stub.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * @generate-class-entries
+ */
+
+final class Mix_Chunk {}
+
