+++ src/Mix_Chunk_arginfo.h
@@ -1,0 +1,20 @@
+/* This is a generated file, edit the .stub.php file instead.
+ * Stub hash: 4d3739c322fe06235f1b4d21c5ad9a8b3bf45f02 */
+
+
+
+
+static const zend_function_entry class_Mix_Chunk_methods[] = {
+	ZEND_FE_END
+};
+
+static zend_class_entry *register_class_Mix_Chunk(void)
+{
+	zend_class_entry ce, *class_entry;
+
+	INIT_CLASS_ENTRY(ce, "Mix_Chunk", class_Mix_Chunk_methods);
+	class_entry = zend_register_internal_class_ex(&ce, NULL);
+	class_entry->ce_flags |= ZEND_ACC_FINAL;
+
+	return class_entry;
+}
