+++ src/Mix_Music.stub.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * @generate-class-entries
+ */
+
+final class Mix_Music {}
+
